pyproject.toml
[project]
name = "household-electricity-dashboard"
version = "1.0.0"
description = "Household Electricity Consumption Dashboard with Tuya Integration"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = [
    "Django>=4.0.0",
    "django-crispy-forms>=1.14.0",
    "Pillow>=9.0.0",
    "celery>=5.2.0",
    "redis>=4.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.26.0",
    "twilio>=7.10.0",
    "pytz>=2021.3",
    "reportlab>=3.6.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "mysqlclient>=2.1.0",
]

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["dashboard", "household_electricity_dashboard"]

[tool.setuptools.package-data]
dashboard = ["templates/**/*.html", "static/**/*", "fixtures/*.json"]

[tool.setuptools.entry-points]
app = { main = "api:application" }

[tool.vercel]
functions = { "api.py" = { maxDuration = 30 } }